Extract ToImmutable
